# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ConfigIn": {
        "properties": {
          "apis": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Apis",
            "type": "array"
          }
        },
        "required": [
          "apis"
        ],
        "title": "ConfigIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LegacyQueryIn": {
        "properties": {
          "nib": {
            "title": "Nib",
            "type": "string"
          }
        },
        "required": [
          "nib"
        ],
        "title": "LegacyQueryIn",
        "type": "object"
      },
      "MappingEditIn": {
        "properties": {
          "concept": {
            "title": "Concept",
            "type": "string"
          },
          "new_field": {
            "title": "New Field",
            "type": "string"
          }
        },
        "required": [
          "concept",
          "new_field"
        ],
        "title": "MappingEditIn",
        "type": "object"
      },
      "QueryIn": {
        "properties": {
          "fields": {
            "title": "Fields",
            "type": "string"
          },
          "input": {
            "title": "Input",
            "type": "string"
          }
        },
        "required": [
          "input",
          "fields"
        ],
        "title": "QueryIn",
        "type": "object"
      },
      "RepositoryConnectIn": {
        "properties": {
          "repository_url": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Repository Url",
            "type": "string"
          }
        },
        "required": [
          "repository_url"
        ],
        "title": "RepositoryConnectIn",
        "type": "object"
      },
      "ResetIn": {
        "properties": {
          "actor": {
            "default": "demo_operator",
            "maxLength": 100,
            "title": "Actor",
            "type": "string"
          }
        },
        "title": "ResetIn",
        "type": "object"
      },
      "ResolveIn": {
        "properties": {
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/SubjectIn"
          }
        },
        "required": [
          "subject",
          "fields",
          "purpose"
        ],
        "title": "ResolveIn",
        "type": "object"
      },
      "SubjectIn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "name",
          "phone"
        ],
        "title": "SubjectIn",
        "type": "object"
      },
      "TestCase": {
        "properties": {
          "expect": {
            "items": {
              "type": "string"
            },
            "title": "Expect",
            "type": "array"
          },
          "expect_values": {
            "additionalProperties": true,
            "title": "Expect Values",
            "type": "object"
          },
          "fields": {
            "title": "Fields",
            "type": "string"
          },
          "forbid": {
            "items": {
              "type": "string"
            },
            "title": "Forbid",
            "type": "array"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "input",
          "fields"
        ],
        "title": "TestCase",
        "type": "object"
      },
      "TestSuiteIn": {
        "properties": {
          "tests": {
            "items": {
              "$ref": "#/components/schemas/TestCase"
            },
            "title": "Tests",
            "type": "array"
          }
        },
        "required": [
          "tests"
        ],
        "title": "TestSuiteIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DataMaster",
    "version": "0.1.0",
    "x-generated-by": "GENERATED FILE — do not edit by hand. Produced by scripts/gen_openapi.py from the route definitions in app/main.py."
  },
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "index__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Index"
      }
    },
    "/api/changes": {
      "get": {
        "operationId": "get_changes_api_changes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Changes"
      }
    },
    "/api/changes/simulate-oss-v2": {
      "post": {
        "description": "Reliable local stand-in for the production GitHub App delivery path.",
        "operationId": "simulate_oss_v2_api_changes_simulate_oss_v2_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Simulate Oss V2"
      }
    },
    "/api/changes/{proposal_id}/approve": {
      "post": {
        "operationId": "post_change_approve_api_changes__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Change Approve"
      }
    },
    "/api/changes/{proposal_id}/mapping": {
      "patch": {
        "operationId": "patch_change_mapping_api_changes__proposal_id__mapping_patch",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MappingEditIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Change Mapping"
      }
    },
    "/api/changes/{proposal_id}/reject": {
      "post": {
        "operationId": "post_change_reject_api_changes__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Change Reject"
      }
    },
    "/api/config": {
      "get": {
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config"
      },
      "put": {
        "operationId": "put_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Config"
      }
    },
    "/api/demo/reset": {
      "post": {
        "operationId": "reset_demo_api_demo_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Demo"
      }
    },
    "/api/demo/tests": {
      "post": {
        "description": "Run isolated acceptance checks without changing the live demo state.",
        "operationId": "demo_tests_api_demo_tests_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Demo Tests"
      }
    },
    "/api/github/webhook": {
      "post": {
        "description": "Model a signed GitHub App event; never apply a proposal automatically.",
        "operationId": "github_webhook_api_github_webhook_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Github Webhook"
      }
    },
    "/api/onboarding": {
      "get": {
        "operationId": "get_onboarding_api_onboarding_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Onboarding"
      }
    },
    "/api/onboarding/query": {
      "post": {
        "operationId": "post_onboarding_query_api_onboarding_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyQueryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Onboarding Query"
      }
    },
    "/api/onboarding/scan": {
      "post": {
        "operationId": "post_onboarding_scan_api_onboarding_scan_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Post Onboarding Scan"
      }
    },
    "/api/onboarding/{proposal_id}/approve": {
      "post": {
        "operationId": "post_onboarding_approve_api_onboarding__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Onboarding Approve"
      }
    },
    "/api/overview": {
      "get": {
        "operationId": "overview_api_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Overview"
      }
    },
    "/api/query": {
      "post": {
        "operationId": "post_query_api_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Query"
      }
    },
    "/api/query/stream": {
      "get": {
        "operationId": "stream_query_api_query_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "input",
            "required": true,
            "schema": {
              "title": "Input",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fields",
            "required": true,
            "schema": {
              "title": "Fields",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Query"
      }
    },
    "/api/repositories/connect": {
      "post": {
        "description": "Document the exposed APIs in a bounded public GitHub repository clone.",
        "operationId": "connect_public_repository_api_repositories_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryConnectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Public Repository"
      }
    },
    "/api/resolve": {
      "post": {
        "description": "Stable downstream contract; callers never send upstream field names.",
        "operationId": "post_resolve_api_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Resolve"
      }
    },
    "/api/services": {
      "get": {
        "operationId": "services_api_services_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Services"
      }
    },
    "/api/tests/run": {
      "post": {
        "operationId": "run_tests_api_tests_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSuiteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tests"
      }
    },
    "/permit": {
      "get": {
        "operationId": "permit_page_permit_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Permit Page"
      }
    }
  }
}